backend/pyproject.toml
[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = "cloudinary"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cloudinary.api"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
env_files = ".env.test"

[tool.poetry]
name = "realtyhub"
version = "0.1.0"
description = ""
authors = ["Stas Maslov <stasyand330@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
fastapi = { extras = ["all"], version = "^0.115.0" }
pyjwt = "2.9.0"
bcrypt = "^4.2.0"
cloudinary = "^1.41.0"
python-multipart = "^0.0.9"
aiosqlite = "^0.20.0"


[tool.poetry.group.test.dependencies]
faker = "^27.0.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
pytest-dotenv = "^0.5.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
ruff = "^0.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
